refactor: listing examples
